requirements.txt
# Core orchestration & FIRMS / weather HTTP clients
requests>=2.31.0
python-dotenv>=1.0.1
pandas>=2.1.0
numpy>=1.26.0
shapely>=2.0.0
pyproj>=3.6.0
scipy>=1.11.0
alphashape>=1.3.1

# Phase 1–2 raster + STAC (Planetary Computer / Sentinel)
pystac-client>=0.7.6
planetary-computer>=1.0.0
stackstac>=0.5.0
xarray>=2024.1.0
rioxarray>=0.15.0
rasterio>=1.3.9
dask[array]>=2024.1.0
